requirement.txt
# conda create -n DiffusionModelFamily python=3.10

# torch
torch==2.3.1
torchvision==0.18.1

# torch util
einops==0.8.0

# diffusers
accelerate==1.2.1
diffusers==0.32.1
transformers==4.47.1
sentencepiece==0.2.0
protobuf==5.29.2
safetensors==0.4.5

# Hydra
hydra-core==1.3.2
omegaconf==2.3.0
pickle5==0.0.11
PyYAML==6.0.2

# misc
# numpy==1.26.2
numpy==2.2.1
matplotlib==3.9.2
pandas==2.2.3
Pillow==10.1.0
tqdm==4.67.1
requests==2.32.3
